Remove users routes
